+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Text based controls**.
   - _$initValue + click on button[clear] + Ctrl+Z => history undo does't work_
   - _button[clear] isn't updated on `$initValue` or `$options.clearActions` changing_
+- [SelectControl](src/controls/select.ts). _Fix clearing input + click outside doesn't clear value_
 
 **New/Features**: